refactor(sync): enforce planner and journal boundaries
